+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- Add `salmon.util` functions `full-name`, `tags`, and `resource`.
+  These are useful when building CloudFormation templates.
 ## v0.12.0 (2023-11-01)
 
 - Add `:client` and `:region` options to stack and stack-properties